# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Rocket Tasks
  description: |
    A fullstack task management API with authentication, lists and tasks.

    ## Authentication
    Most endpoints require a Bearer token. Use `POST /auth/login` or `POST /auth/register` to obtain a token, then include it in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: Rocket Tasks
    url: https://github.com/your-username/rocket-tasks
security: []
...
paths:
...
  /users/me:
    get:
      summary: Get Current User
      operationId: getMe
      description: Returns the profile information of the authenticated user.
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 60af62a8-3384-402e-a3e4-203b37cf183b
                name: John Doe
                email: john@example.com
                avatarUrl: null
                createdAt: '2026-07-01T10:00:00.000Z'
                updatedAt: '2026-07-01T10:00:00.000Z'
                _count:
                  tasks: 5
                  lists: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Update Current User
      operationId: updateMe
      description: Updates the authenticated user's profile and returns the updated user information.
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/update_current_user'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 60af62a8-3384-402e-a3e4-203b37cf183b
                name: John Updated
                email: john@example.com
                avatarUrl: null
                createdAt: '2026-07-01T10:00:00.000Z'
                updatedAt: '2026-07-29T10:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete Account
      operationId: deleteMe
      description: Permanently deletes the authenticated user's account and all associated resources.
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            tasks:
              type: integer
            lists:
              type: integer
...
    Error:
      type: object
      properties:
        message:
          type: string
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              path:
                type: array
                items:
                  type: string
...
    update_current_user:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
      example:
        name: John Updated
...
  requestBodies:
...
    update_current_user:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/update_current_user'
      required: true
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Token not provided
...
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            errors:
              - code: invalid_format
                message: Invalid email address
                path:
                  - email
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Internal server error